requirements.txt
python-telegram-bot==21.0
openai==1.54.0
pyyaml==6.0.2
python-dotenv==1.0.1
aiosqlite==0.20.0
